PR suggestions for modularity and code reduction
