+++ nwbinspector/load_checks.py
@@ -1,6 +1,4 @@
 """Authors: Cody Baker and Ben Dichter."""
-from copy import deepcopy
-
 from .utils import default_checks
 from .check_time_series import (
     check_dataset_compression,
@@ -9,5 +7,5 @@ from .check_time_series import (
 )
 
 
-def load_checks(checks=deepcopy(default_checks)):
+def load_checks(checks=default_checks):
     return default_checks
